RAG-System/requirements.txt
openai
python-dotenv
pypdf
tiktoken
pandas
numpy
pinecone
pinecone[grpc]
pyyaml